requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
billiard==3.6.4.0
branca==0.6.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
cryptography==38.0.4
daphne==4.0.0
Django==4.1.3
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-crontab==0.7.1
django-queryset-csv==1.1.0
django-rest-framework==0.1.0
django-sslserver==0.22
django-timezone-field==5.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
folium==0.14.0
geojson==2.5.0
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
msgpack==1.0.4
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.2
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.6
redis==4.4.0
rel==0.4.7
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0

txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
unicodecsv==0.14.1
urllib3==1.26.13
uvicorn==0.20.0
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.4.2
websockets==10.4
zope.event==4.6
zope.interface==5.5.2
